feat(event): 将群组相关事件中的 uint 类型字段升级为 ulong

+++ Lagrange.Core/Event/EventArg/GroupAdminChangedEvent.cs
@@ -2,13 +2,13 @@ namespace Lagrange.Core.Event.EventArg;
 
 public class GroupAdminChangedEvent : EventBase
 {
-    public uint GroupUin { get; }
+    public ulong GroupUin { get; }
     
-    public uint AdminUin { get; }
+    public ulong AdminUin { get; }
     
     public bool IsPromote { get; }
     
-    public GroupAdminChangedEvent(uint groupUin, uint adminUin, bool isPromote)
+    public GroupAdminChangedEvent(ulong groupUin, ulong adminUin, bool isPromote)
     {
         GroupUin = groupUin;
         AdminUin = adminUin;
