+++ Lagrange.Core/Event/EventArg/GroupMuteEvent.cs
@@ -2,13 +2,13 @@ namespace Lagrange.Core.Event.EventArg;
 
 public class GroupMuteEvent : EventBase
 {
-    public uint GroupUin { get; }
+    public ulong GroupUin { get; }
     
     public uint? OperatorUin { get; }
     
     public bool IsMuted { get; }
     
-    public GroupMuteEvent(uint groupUin, uint? operatorUin, bool isMuted)
+    public GroupMuteEvent(ulong groupUin, uint? operatorUin, bool isMuted)
     {
         GroupUin = groupUin;
         OperatorUin = operatorUin;
