+++ Lagrange.Core/Internal/Event/Notify/GroupSysChangeName.cs
@@ -2,16 +2,16 @@ namespace Lagrange.Core.Internal.Event.Notify;
 
 internal class GroupSysNameChangeEvent : ProtocolEvent
 {
-    public uint GroupUin { get; }
+    public ulong GroupUin { get; }
 
     public string Name { get; }
 
-    private GroupSysNameChangeEvent(uint groupUin, string name) : base(0)
+    private GroupSysNameChangeEvent(ulong groupUin, string name) : base(0)
     {
         GroupUin = groupUin;
         Name = name;
     }
 
-    public static GroupSysNameChangeEvent Result(uint groupUin, string name)
+    public static GroupSysNameChangeEvent Result(ulong groupUin, string name)
         => new(groupUin, name);
 }
